+++ src/data_types.py
@@ -18,6 +18,15 @@ UpscalerInfo = TypedDict(
         "name": str,
     },
 )
+TextualInversionInfo = TypedDict(
+    "TextualInversionInfo",
+    {
+        "file": str,
+        "keyword": str,
+        "name": str,
+        "version": str,
+    }
+)
 ModelsInfo = TypedDict(
     "ModelsInfo",
     {
@@ -25,6 +34,7 @@ ModelsInfo = TypedDict(
         "controlNets": list[ControlNetInfo],
         "loras": list[LoRAInfo],
         "upscalers": list[UpscalerInfo],
+        "textualInversions": list[TextualInversionInfo],
     },
 )
